fix: move the remaining data from styled-components to css modules
