+++ src/components/Card/Cards.module.css
@@ -1,0 +1,16 @@
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  gap: var(--spacings-s);
+  margin-bottom: var(--spacings-s);
+}
+
+.footer {
+  display: flex;
+  gap: var(--spacings-xxs);
+  justify-content: center;
+
+  @media only screen and (min-width: 1200px) {
+    justify-content: flex-start;
+  }
+}
