+++ src/components/Label/Label.module.css
@@ -1,0 +1,6 @@
+.label {
+  padding: var(--spacings-xxs);
+  width: max-content;
+  list-style-type: none;
+  font-size: 0.8rem;
+}
