+++ src/components/Main/Main.module.css
@@ -1,0 +1,51 @@
+.container {
+  padding: var(--spacings-xxs) var(--spacings-s);
+  display: flex;
+  gap: var(--spacings-s);
+  flex-wrap: wrap;
+
+  @media only screen and (min-width: 1200px) {
+    flex-wrap: nowrap;
+    padding: var(--spacings-s) var(--spacings-m);
+    gap: var(--spacings-m);
+  }
+}
+
+.copy {
+  font-size: 1.2rem;
+  max-width: 100%;
+
+  @media only screen and (min-width: 1200px) {
+    max-width: 40%;
+  }
+}
+
+.copy h1 {
+  margin-bottom: var(--spacings-xs);
+  font-size: 2.5rem;
+}
+
+.copy h3 {
+  font-size: 1.2rem;
+}
+
+.copy p {
+  line-height: 2.2rem;
+}
+
+.sticky {
+  position: sticky;
+  top: var(--spacings-m);
+}
+
+.skills {
+  display: flex;
+  gap: var(--spacings-xs);
+  padding: 0;
+  align-items: center;
+  flex-wrap: wrap;
+}
+
+.copy p > .wttLink {
+  border-bottom: 2.5px dashed var(--palette-dark-orange);
+}
