+++ src/components/Main/Main.tsx
@@ -1,39 +1,35 @@
 import React from "react";
-import { Container, Copy, Description, Sticky, Skills } from "./Main.styled";
-import Cards from "../Card/Card";
+import Cards from "../Card/Cards";
 import Label from "../Label/Label";
 import { MainData, Project, TechStack } from "../interfaces";
 import { skills } from "../../backups/skills";
+import styles from "./Main.module.css";
 
 interface MainProps extends MainData {
   items: Project[];
 }
 
-const Main = ({ header, description, tags, items }: MainProps) => {
-  return (
-    <Container>
-      <Copy>
-        <Sticky>
-          <h1 dangerouslySetInnerHTML={{ __html: header }} />
-          <Description
-            dangerouslySetInnerHTML={{ __html: description }}
-          ></Description>
-          <article>
-            <h2>Skills</h2>
-            {Object.keys(tags || skills).map((name) => (
-              <Skills key={name}>
-                <h3>{name}</h3>
-                {tags[name].map((item: TechStack, i: number) => (
-                  <Label text={item} key={i} />
-                ))}
-              </Skills>
-            ))}
-          </article>
-        </Sticky>
-      </Copy>
-      <Cards items={items} />
-    </Container>
-  );
-};
+const Main = ({ header, description, tags, items }: MainProps) => (
+  <main className={styles.container}>
+    <section className={styles.copy}>
+      <div className={styles.sticky}>
+        <h1 dangerouslySetInnerHTML={{ __html: header }} />
+        <p dangerouslySetInnerHTML={{ __html: description }} />
+        <article>
+          <h2>Skills</h2>
+          {Object.keys(tags || skills).map((name) => (
+            <ul key={name} className={styles.skills}>
+              <h3>{name}</h3>
+              {tags[name].map((item: TechStack, i: number) => (
+                <Label text={item} key={i} />
+              ))}
+            </ul>
+          ))}
+        </article>
+      </div>
+    </section>
+    <Cards items={items} />
+  </main>
+);
 
 export default Main;
